+++ test/irmin-mem/test_mem.ml
@@ -23,9 +23,18 @@ let config = Irmin_mem.config ()
 
 let clean () =
   let (module S : Irmin_test.S) = store in
+  let module P = S.Private in
+  let clear repo =
+    Lwt.join
+      [
+        P.Commit.clear (P.Repo.commit_t repo);
+        P.Node.clear (P.Repo.node_t repo);
+        P.Contents.clear (P.Repo.contents_t repo);
+        P.Branch.clear (P.Repo.branch_t repo);
+      ]
+  in
   S.Repo.v config >>= fun repo ->
-  S.Repo.branches repo >>= Lwt_list.iter_p (S.Branch.remove repo) >>= fun () ->
-  S.Repo.close repo
+  clear repo >>= fun () -> S.Repo.close repo
 
 let init () = Lwt.return_unit